Move context action and selector to private
