+++ packages/block-editor/src/store/private-actions.js
@@ -415,3 +415,19 @@ export const modifyContentLockBlock =
 			focusModeToRevert
 		);
 	};
+
+/**
+ * Action that changes the block context of the given block in the store.
+ *
+ * @param {string } clientId Client ID of the block.
+ * @param {Object}  context  Object with the new context.
+ *
+ * @return {Object} Action object
+ */
+export function updateBlockContext( clientId, context ) {
+	return {
+		type: 'UPDATE_BLOCK_CONTEXT',
+		clientId,
+		context,
+	};
+}
